+++ cocos/editor-support/spine/Atlas.cpp
@@ -142,7 +142,7 @@ static int readTuple (const char* end, Str tuple[]) {
 }
 
 static char* mallocString (Str* str) {
-	int length = str->end - str->begin;
+	long length = str->end - str->begin;
 	char* string = MALLOC(char, length + 1);
 	memcpy(string, str->begin, length);
 	string[length] = '\0';
@@ -150,7 +150,7 @@ static char* mallocString (Str* str) {
 }
 
 static int indexOf (const char** array, int count, Str* str) {
-	int length = str->end - str->begin;
+	long length = str->end - str->begin;
 	int i;
 	for (i = count - 1; i >= 0; i--)
 		if (strncmp(array[i], str->begin, length) == 0) return i;
@@ -162,7 +162,7 @@ static int equals (Str* str, const char* other) {
 }
 
 static int toInt (Str* str) {
-	return strtol(str->begin, (char**)&str->end, 10);
+	return static_cast<int>(strtol(str->begin, (char**)&str->end, 10));
 }
 
 static spAtlas* abortAtlas (spAtlas* self) {
@@ -177,7 +177,7 @@ static const char* textureFilterNames[] = {"Nearest", "Linear", "MipMap", "MipMa
 spAtlas* spAtlas_readAtlas (const char* begin, int length, const char* dir) {
 	int count;
 	const char* end = begin + length;
-	int dirLength = strlen(dir);
+	size_t dirLength = strlen(dir);
 	int needsSlash = dirLength > 0 && dir[dirLength - 1] != '/' && dir[dirLength - 1] != '\\';
 
 	spAtlas* self = NEW(spAtlas);
@@ -289,7 +289,7 @@ spAtlas* spAtlas_readAtlas (const char* begin, int length, const char* dir) {
 }
 
 spAtlas* spAtlas_readAtlasFile (const char* path) {
-	int dirLength;
+	long dirLength;
 	char *dir;
 	int length;
 	const char* data;
